+++ source/danger.d.ts
@@ -832,6 +832,8 @@ interface TextDiff {
 interface StructuredDiff {
   /** Git diff chunks */
   chunks: File["chunks"]
+  /** The file path pre-change */
+  fromPath: string | undefined
 }
 
 /** The results of running a JSON patch */